Addressing more review feedback.
- updated and moved part of the logic for whether or not to include the files needed for the modal.
- use the version attribute already provided by the `*.asset.php` file in the build directory
- instantiate the new controller immediately after including the file instead of directly in the file itself.
- Renamed the class to be more specific to its function.
- Moved the new component files into the `client/entrypoints` folder.
