+++ client/entrypoints/payment-gateways/payment-gateways-confirmation.js
