+++ client/entrypoints/payment-gateways/style.scss
