+++ webpack.config.js
@@ -40,6 +40,6 @@ module.exports = {
 		upe_onboarding_wizard: './client/upe-onboarding-wizard/index.js',
 		upe_opt_in_banner: './client/entrypoints/upe-opt-in-banner/index.js',
 		upe_settings: './client/settings/index.js',
-		payment_gateways: './client/payment-gateways/index.js',
+		payment_gateways: './client/entrypoints/payment-gateways/index.js',
 	},
 };
